+++ client/plots/profileRadarFacility.js
@@ -54,7 +54,7 @@ class profileRadarFacility extends profilePlot {
 	plot() {
 		this.dom.plotDiv.selectAll('*').remove()
 		if (this.data.lst.length == 0) return
-		const widht = 1200
+		const widht = 1100
 		const height = 650
 		this.svg = this.dom.plotDiv
 			.append('div')
@@ -80,11 +80,11 @@ class profileRadarFacility extends profilePlot {
 		const polarG = this.svg.append('g').attr('transform', `translate(${x},${y})`)
 		this.polarG = polarG
 
-		this.legendG = this.svg.append('g').attr('transform', `translate(${x + 350},${y + 40})`)
-		this.filterG = this.svg.append('g').attr('transform', `translate(${x + 350},${y + 140})`)
+		this.legendG = this.svg.append('g').attr('transform', `translate(${x + 330},${y + 100})`)
+		this.filterG = this.svg.append('g').attr('transform', `translate(${x + 330},${y + 200})`)
 
 		const rows = []
-		const columns = [{ label: 'Color' }, { label: 'Module' }, { label: `Facility` }, { label: 'Global' }]
+		const columns = [{ label: 'Module' }, { label: `Facility` }, { label: 'Global' }]
 
 		for (let i = 0; i <= 10; i++) this.addPoligon(i * 10)
 
@@ -94,8 +94,7 @@ class profileRadarFacility extends profilePlot {
 		for (const item of this.terms) {
 			const iangle = i * this.angle - Math.PI / 2
 			this.addData(iangle, i, data2, true)
-			const color = item.score.term.color
-			const row = [{ color }, { value: item.module }, { value: this.getPercentage(item) }]
+			const row = [{ value: item.module }, { value: this.getPercentage(item) }]
 
 			this.addData(iangle, i, data, false)
 			row.push({ value: this.getPercentage2(item) })
@@ -230,6 +229,7 @@ class profileRadarFacility extends profilePlot {
 
 		const textElem = this.legendG
 			.append('text')
+			.attr('font-size', '0.9em')
 			.attr('transform', `translate(${x + 5}, ${y})`)
 			.attr('text-anchor', 'left')
 		textElem.append('tspan').text(text)