+++ client/plots/sampleScatter.js
@@ -382,7 +382,7 @@ class Scatter {
 		}
 		// TODO: handle multiple chart download when there is a divide by term
 		this.components.controls.on('downloadClick.scatter', () => {
-			for (const chart of this.charts) downloadSingleSVG(chart.svg, 'scatter.svg')
+			for (const chart of this.charts) downloadSingleSVG(chart.svg, 'scatter.svg', this.opts.holder.node())
 		})
 		this.dom.toolsDiv = this.dom.controls.insert('div')
 	}