+++ server/src/termdb.config.js
@@ -46,10 +46,12 @@ export function make(q, res, ds, genome) {
 		massSessionDuration: serverconfig.features.massSessionDuration || 30,
 		dataDownloadCatch: tdb.dataDownloadCatch,
 		matrix: tdb.matrix,
+		hierCluster: tdb.hierCluster,
 		mclass: tdb.mclass,
 		alwaysRefillCategoricalTermValues: tdb.alwaysRefillCategoricalTermValues,
 		isGeneSetTermdb: tdb.isGeneSetTermdb,
-		lollipop: tdb.lollipop
+		lollipop: tdb.lollipop,
+		urlTemplates: tdb.urlTemplates
 	}
 	// optional attributes
 	// when missing, the attribute will not be present as "key:undefined"