+++ server/src/termdb.matrix.js
@@ -108,6 +108,8 @@ async function getSampleData(q) {
 		// for each non dictionary term type
 		// query sample data with its own method and append results to "samples"
 		if (tw.term.type == 'geneVariant') {
+			if (q.ds.cohort?.termdb?.getGeneAlias) refs.byTermId[tw.term.name] = q.ds.cohort?.termdb?.getGeneAlias(q, tw)
+
 			const bySampleId = await q.ds.mayGetGeneVariantData(tw, q)
 			for (const [sampleId, value] of bySampleId.entries()) {
 				if (!(tw.term.name in value)) continue